refactor(p1): extract Group C SystemAccess domains — Apps/DeviceStatus/Intent/Share/NotificationPost
Final extraction round for SystemAccess. Closes the god-class split:
SystemAccessTools.kt 1441 → 93 lines (-93%, pure coordinator now).

Four new sibling files + one extension to NotificationsAccessTools:
- AppsAccessTools.kt — createAppsListTool / createAppOpenTool /
  createAppsInstalledListTool / createAppInfoTool plus three file-private
  helpers (queryLaunchableApps, queryInstalledApps, packageInfoJson).
- DeviceStatusAccessTools.kt — battery / network / wifi / device-info
  factories plus four `XxxJson(context)` helpers; redactMiddle moves here
  file-private (sole caller is wifiStatusJson). IntentFilter import added
  (was used inline as `android.content.IntentFilter` in the prior fully-
  qualified form).
- IntentAccessTools.kt — settings_open / intent_open factories plus
  settingsIntent(context, target) and whitelistedIntent (Uri scheme
  whitelist preserved verbatim — http/https/tel/mailto/smsto).
- ShareAccessTools.kt — share_text + share_file factories. share_file's
  createShareFileTool takes workspaceManager as a third explicit param
  (the only tool that needs it); cacheWorkspaceFileForSharing and
  mimeFromExtension file-private. MimeTypeMap import added (was fully-
  qualified before).
- NotificationsAccessTools.kt — extended with createNotificationPostTool
  alongside the existing createNotificationListTool (both are conceptually
  "notifications" tooling).

SystemAccessTools.kt becomes a thin coordinator: 4-dep ctor, deps wrapper
field, 28 `by lazy { createXxxTool(...) }` delegators, getTools() registry
(preserved tool order intact). Class doc-comment added pointing readers
to SystemAccessShared.kt for the cross-domain infra.

Drops 26 now-dead imports from SystemAccessTools.kt: NotificationManager,
PendingIntent, UiModeManager, ActivityNotFoundException, Intent,
ApplicationInfo, PackageManager, ConnectivityManager, NetworkCapabilities,
Uri, WifiManager, Build, BatteryManager, PowerManager, Settings,
NotificationCompat, FileProvider, JsonElement, add, buildJsonArray,
buildJsonObject, put, CHAT_COMPLETED_NOTIFICATION_CHANNEL_ID, R, File.
Only Context, Tool, the three dep types, and WorkspaceManager remain.

Cumulative result (Sessions #3+#4 P1 god-class work):
- LocalTools.kt: 1093 → 189 (Session #3, 14 sibling files, -83%)
- SystemAccessTools.kt: 1441 → 93 (Session #4, 13 sibling files, -93%)

Verification: :app:compileDebugKotlin --offline → BUILD SUCCESSFUL.
